requirements.txt
langchain
langchain-huggingface
numpy
pandas
nltk
streamlit
python-dotenv
transformers
torch
scikit-learn